backend/requirements.txt
flask==2.0.1
flask-babel==2.0.0
boto3==1.26.137
zappa==0.54.0
python-dotenv==0.19.0 